fea-35 Refactor SubscribeBubbles and Story to TS
